+++ src/aosm/HISTORY.rst
@@ -8,6 +8,8 @@ Unreleased
 * Add `publisher` command group for management of publisher resources.
 * Changed the name of the `path_to_mappings` parameter in the CNF input file to `default_values`
 * Added a `helm template` validation step to the `az aosm nfd build` command for the `cnf` definition type
+* Fixed helm chart image parsing in the `az aosm nfd build` command for the `cnf` definition type. This means that the images can now be extracted correctly from the helm chart.
+* Fixed: infinite loop bug when retrying failed artifact uploads to the ACR
 
 1.0.0b4
 ++++++++